Signup
- Added the validations;
- Added the top messages view;

TODO:
- Adjust for the 4S
- The drag to dimiss the top view is not correct
- Birthday fields is allowing the user to provide any date;

+++ mugchat/mugchat/Classes/UIImage+ImageEffects.m
@@ -127,7 +127,7 @@
 {
     const CGFloat EffectColorAlpha = 0.9;
     UIColor *effectColor = tintColor;
-    int componentCount = CGColorGetNumberOfComponents(tintColor.CGColor);
+    long componentCount = CGColorGetNumberOfComponents(tintColor.CGColor);
     if (componentCount == 2) {
         CGFloat b;
         if ([tintColor getWhite:&b alpha:NULL]) {
